% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qst.R
\docType{package}
\name{qst}
\alias{qst}
\alias{qst-package}
\title{Store Tables in SQL Database}
\description{
This package provides functions for quickly writing (and reading)
back a \code{data.frame} to file in \code{sqlite} format. The name stands
for \emph{Store Tables using SQLite'}, or alternatively for \emph{Quick
Store Tables} (either way, it could be pronounced as \emph{Quest}).

For \code{data.frames} containing the supported data
types it is intended to work as a drop-in replacement for the
\verb{write_*()} and \verb{read_*()} functions provided by packages such
as \code{fst}, \code{feather}, \code{qs}, and \code{readr} packages (as well as the
\code{writeRDS()} and \code{readRDS()} functions).
}
